bulk of the changes
