+++ frontend/src/App.tsx
@@ -1,9 +1,9 @@
-import React, { ReactChildren } from 'react'
+import React, { FunctionComponent } from 'react'
 import {
   Landing,
-  // LoginForm,
-  // RegisterForm,
-  // Menu,
+  LoginForm,
+  RegisterForm,
+  Menu,
   // Surveys,
   Candidacies,
 } from './Pages'
@@ -11,20 +11,22 @@ import { Switch, Route, Redirect } from 'react-router-dom'
 import { useSelector, useDispatch } from 'react-redux'
 import { Container } from 'react-bootstrap'
 
-function App() {
+const App: FunctionComponent = () => {
   const dispatch = useDispatch()
   dispatch({ type: 'LOAD_FROM_COOKIES' })
 
   const isLoggedIn = useSelector((state: any) => state.login.isLoggedIn)
 
-  interface test {
+  interface RouteInterface {
     children: any,
+    exact: Boolean,
     path: string
   }
 
-  const AuthorizedRoute: React.FC<test> = ({ children, path }) => {
+  const AuthorizedRoute: React.FC<RouteInterface> = ({ children, path }) => {
     return (
       <Route
+        exact
         path={path}
         render={({ location }) =>
           isLoggedIn ? (
@@ -40,7 +42,7 @@ function App() {
     )
   }
 
-  const NoAuthorizationRoute: React.FC<test> = ({ children, path}) => {
+  const NoAuthorizationRoute: React.FC<RouteInterface> = ({ children, path }) => {
     return (
       <Route
         path={path}
@@ -61,22 +63,22 @@ function App() {
   return (
     <Container fluid className="appContainer">
       <Switch>
-        <NoAuthorizationRoute /*exact*/ path="/">
+        <NoAuthorizationRoute exact path="/">
           <Landing />
         </NoAuthorizationRoute>
-        <NoAuthorizationRoute /*exact*/ path="/login">
-          {/* <LoginForm /> */}
+        <NoAuthorizationRoute exact path="/login">
+          <LoginForm />
         </NoAuthorizationRoute>
-        <NoAuthorizationRoute /*exact*/ path="/register">
-          {/* <RegisterForm /> */}
+        <NoAuthorizationRoute exact path="/register">
+          <RegisterForm />
         </NoAuthorizationRoute>
-        <AuthorizedRoute /*exact*/ path="/menu">
-          {/* <Menu /> */}
+        <AuthorizedRoute exact path="/menu">
+          <Menu />
         </AuthorizedRoute>
-        <AuthorizedRoute path="/surveys">
+        <AuthorizedRoute exact={false} path="/surveys">
           {/* <Surveys /> */}
         </AuthorizedRoute>
-        <AuthorizedRoute /*exact*/ path="/candidacies">
+        <AuthorizedRoute exact path="/candidacies">
           {/* <Candidacies /> */}
         </AuthorizedRoute>
         <Route path="*">
