+++ frontend/src/Pages/index.ts
@@ -1,5 +1,6 @@
 export * from './Landing/landing'
-export * from './LoginRegister/loginRegister'
+export * from './LoginRegister/login'
+export * from './LoginRegister/register'
 export * from './Menu/menu'
 export * from './Surveys/surveySelection'
 export * from './Candidacies/candidacies'
